+++ mopidy/frontends/mpd/protocol/playback.py
@@ -293,6 +293,8 @@ def seek(frontend, songpos, seconds):
         Seeks to the position ``TIME`` (in seconds) of entry ``SONGPOS`` in
         the playlist.
     """
+    if frontend.backend.playback.current_playlist_position != songpos:
+        playpos(frontend, songpos)
     return frontend.backend.playback.seek(int(seconds) * 1000)
 
 @handle_pattern(r'^seekid "(?P<cpid>\d+)" "(?P<seconds>\d+)"$')
@@ -304,7 +306,9 @@ def seekid(frontend, cpid, seconds):
 
         Seeks to the position ``TIME`` (in seconds) of song ``SONGID``.
     """
-    raise MpdNotImplemented # TODO
+    if frontend.backend.playback.current_cpid != cpid:
+        playid(frontend, cpid)
+    return frontend.backend.playback.seek(int(seconds) * 1000)
 
 @handle_pattern(r'^setvol "(?P<volume>[-+]*\d+)"$')
 def setvol(frontend, volume):